Revert "use the proper doc type"
This reverts commit b65b1257f8f5e7c1138f041e0c804b249519a024.

Apparently the reason why this was using doc type html without the dtd is that is the way HTML5 expects.  When I try to validate against xhtml, I get an error that some of the <div> tags have a role attribute that is apparently not introduced till html5.  While my .setup file _says_ html5 is disabled, we still seem to be using at least some of it.