build: check point
a number of improvements to the RecentChanges generation
it better detects which changes have files,
it better creates a list of those files
it still does not successfully find 100 such changes
it still does not do relative dates to now.
it better finds the file metadata.

revisited the ikiwiki conversion, this time in perl
I am more confident of this version, so it actually
removes .mdwn files
it is not quite right yet, I am trying to clean up
from the ancient conversion from wordpress as well
and that is only partly working. It *does*
work better than convert.sh did.
It does not yet handle authors.
It does not yet generate tag pages.
It does not yet generate calender pages.

I may not want it to do either of the last two.
I may want to make the permenently part of the
compile step - for long term maintenance of the site.