build: check point
the App::History module was redundant and buggy, remove it
CalendarGenerator now does most of what I want, still need a few validations
Add to the controller the beginings of rendering nav content
* the start of a sidebar
* some helpers for calendar pages and calendars in sidebars

Add in some css to help with displaying the sidebar as such.
The CSS for the sidebar isn't quite right yet.